+++ CHANGELOG.md
@@ -19,6 +19,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - `TryFrom<TokenTree> for litrs::Literal` impls
 - `From<*Lit> for litrs::Literal` impls (e.g. `From<StringLit> for Literal`)
+- `TryFrom<proc_macro[2]::Literal> for *Lit` (e.g. `for IntegerLit`)
 
 
 ## [0.1.1] - 2021-05-25